1.实现了增删改查套餐的功能 2.对一些sql语句进行了优化 3.现在当种类关联了某个菜品或套餐时，不可以删除了

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -1,15 +1,22 @@
 package com.sky.service.impl;
 
 import cn.hutool.core.bean.BeanUtil;
+import cn.hutool.core.collection.CollUtil;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.sky.constant.MessageConstant;
 import com.sky.constant.StatusConstant;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.entity.Category;
+import com.sky.entity.Dish;
+import com.sky.entity.Setmeal;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.mapper.CategoryMapper;
 import com.sky.result.PageResult;
 import com.sky.service.CategoryService;
+import com.sky.service.DishService;
+import com.sky.service.SetmealService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,6 +27,12 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+
+    @Autowired
+    private DishService dishService;
+
+    @Autowired
+    private SetmealService setmealService;
 
     /**
      * 添加分类
@@ -38,9 +51,9 @@ public class CategoryServiceImpl implements CategoryService {
      * @return
      */
     @Override
-    public PageResult<Category> getCategoryList(CategoryPageQueryDTO categoryPageQueryDTO) {
+    public PageResult<Category> page(CategoryPageQueryDTO categoryPageQueryDTO) {
         PageHelper.startPage(categoryPageQueryDTO.getPage(), categoryPageQueryDTO.getPageSize());
-        Page<Category> page = categoryMapper.getCategoryList(categoryPageQueryDTO.getName(), categoryPageQueryDTO.getType());
+        Page<Category> page = categoryMapper.page(categoryPageQueryDTO.getName(), categoryPageQueryDTO.getType());
         PageResult<Category> pageResult = new PageResult<>(page.getTotal(), page.getResult());
         return pageResult;
     }
@@ -51,6 +64,17 @@ public class CategoryServiceImpl implements CategoryService {
      */
     @Override
     public void deleteCategory(Long id) {
+        // 1.如果分类关联了菜品,不能删除
+        List<Dish> dishList = dishService.getDishesByCategoryId(id);
+        if (CollUtil.isNotEmpty(dishList)){
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+        }
+        // 2.如果分类关联了套餐,不能删除
+        List<Setmeal> setmealList = setmealService.getSetmealsByCategoryId(id);
+        if (CollUtil.isNotEmpty(setmealList)){
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_SETMEAL);
+        }
+        // 3.删除分类
         categoryMapper.deleteCategory(id);
     }
